build(conda): renable python noarch flag
As of version 0.2, the `trollius` package was added as a requirement in
the Conda build recipe to provide asynchronous IO support in Python 2.7
using the [`py27` preprocessing selector][selectors] to only install it in
the Python 2.7 package build.  However, this also required the removal of
the `noarch: python` flag was removed from the Conda build recipe, since
(quoted from [here][noarch]:

> ... noarch packages should not make use of preprocess-selectors: noarch
> packages are built with the directives which evaluate to True in the
> platform it was built, which probably will result in
> incorrect/incomplete installation in other platforms.

In this commit, remove the `py27` preprocessing selector for the
`trollius` package, since the package is available for Python 2.7 **and**
Python 3+ in the `conda-forge` channel.  Also, since no preprocessing
selectors are in use, reintroduce the `noarch: python` flag to the Conda
build recipe.

[selectors]: https://conda.io/docs/user-guide/tasks/build-packages/define-metadata.html#preprocessing-selectors
[noarch]: https://conda.io/docs/user-guide/tasks/build-packages/define-metadata.html#architecture-independent-packages