test(close): add test of closing AsyncSerial
In Windows, it turns out that the serial port is tied up by an AsyncSerial
instance until the corresponding event loop is closed.

In this commit, add a test to demonstrate this and test a workaround,
which is to close the corresponding event loop to free up the serial port.