feat(py3/thread): add BackgroundSerialAsync
IPython/Jupyter kernels running on Python 3 implicitly instantiate an
`asyncio` event loop in the main kernel thread. On Windows, this is a
`SelectorEventLoop`. Unfortunately, `asyncserial` code requires a
`ProactorEventLoop` to work on Windows. This prevents `asyncserial` from
being used on Windows from IPython/Jupyter kernels (e.g., in Jupyter
notebooks).

In this commit, add a new `BackgroundSerialAsync` class which launches
an `AsyncSerial` connection in a background thread with its own
`ProactorEventLoop`. An async API is exposed to transparently redirect
from coroutines in the calling thread to the respective methods in the
background thread, using `asyncio.Event` to coordinate with the calling
thread.

**Note that `BackgroundSerialAsync` is currently implemented for Python
3 only, but could be ported to work with Python 2 using the `trollius`
library.**